Added shared preferences to AppModule.kt for di.
Updated ExerciseViewModel.kt to use  shared preferences to to get/update the weight unit setting.

Added UI to select on PreferenceScreen.kt. Added an enum for WeightUnit as well.

Updated the ViewExerciseScreen.kt to show weight value and unit based on the weight unit setting. Added a extension function for conversion from kg to pounds.

 Added a extension function for conversion from pounds to kg on LogPBScreen.kt.

 Added new strings to strings.xml.